requirements.txt
Flask==3.0.3
Jinja2==3.1.4
mysql-connector-python==9.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1